Terragen: fix CID 1074048, 1074049
git-svn-id: https://svn.osgeo.org/gdal/trunk@31906 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/terragen/terragendataset.cpp
@@ -805,7 +805,7 @@ int TerragenDataset::LoadFromFile()
             get(m_nHeightScale);
             get(m_nBaseHeight);
             m_nDataOffset = VSIFTellL(m_fp);
-            if( !skip(xpts * ypts * sizeof(GInt16)) )
+            if( !skip(static_cast<size_t>(xpts) * static_cast<size_t>(ypts) * sizeof(GInt16)) )
                 return FALSE;
             continue;
         }
